refactor gambit export
