+++ pycid/__init__.py
@@ -1,3 +1,4 @@
 from pycid.analyze import *  # noqa
 from pycid.core import *  # noqa
+from pycid.export import *  # noqa
 from pycid.random import *  # noqa
